// packages/connect/ic/providers/types.ts
export const iiDerivationOrigin = location.origin.includes('easydapp.ai')
    ? 'https://fyhem-pyaaa-aaaai-q3jwa-cai.icp0.io/'
    : undefined;

type CustomError<T> = { kind: T; message?: string };

export enum CreateActorError {
    FetchRootKeyFailed = 'FETCH_ROOT_KEY_FAILED',
    CreateActorFailed = 'CREATE_ACTOR_FAILED',
    NotInitialized = 'NOT_INITIALIZED',
    LocalActorsNotSupported = 'LOCAL_ACTORS_NOT_SUPPORTED',
}

export type CreateActorResult<Service> = Result<
    ActorSubclass<Service>,
    CustomError<CreateActorError>
>;

export enum ConnectError {
    NotInitialized = 'NOT_INITIALIZED',
    NotInstalled = 'NOT_INSTALLED',
    ConnectFailed = 'CONNECT_FAILED',
    IsLocked = 'IS_LOCKED',
}

export type ConnectResult = Result<boolean, CustomError<ConnectError>>;

export enum DisconnectError {
    DisconnectFailed = 'DISCONNECT_FAILED',
    NotInitialized = 'NOT_INITIALIZED',
}

export type DisconnectResult = Result<boolean, CustomError<DisconnectError>>;

export enum InitError {
    NotInstalled = 'NOT_INSTALLED',
    InitFailed = 'INIT_FAILED',
    FetchRootKeyFailed = 'FETCH_ROOT_KEY_FAILED',
}